Site Settings: Move Site Verification in a separate component
